Change CCL input precision from int8 to int64 to align with compiler

+++ QEfficient/generation/vlm_generation.py
@@ -315,7 +315,7 @@ class VisionLanguageGeneration(QEffTextGenerationBase):
 
         if self.comp_ctx_lengths_prefill is not None:
             self.list_of_comp_ctx_lengths_prefill = [
-                np.zeros(length, dtype=np.int8) for length in self.comp_ctx_lengths_prefill
+                np.zeros(length, dtype=np.int64) for length in self.comp_ctx_lengths_prefill
             ]
             prefill_ccl_id = 0
             lang_inputs["comp_ctx_lengths"] = self.list_of_comp_ctx_lengths_prefill[prefill_ccl_id]
